ops/requirements.txt
google-auth==2.17.3
google-api-python-client==2.88.0
pandas>=1.5.0
numpy>=1.21.0
